тип системных переменных изменён на двухбайтный, старт логики отложен до успешной инициализации Ethernet

+++ Src/ld_prog.c
@@ -40,13 +40,13 @@ uint16_t net_regs_tx[16]={0};
 uint16_t prev_net_regs_tx[16]={0};
 
 // SS1..SS8 (nodes online offline)
-extern uint8_t node_link[8];
+extern uint16_t node_link[8];
 
 // SS9 PC21 Network Can
-extern uint8_t can_link;
+extern uint16_t can_link;
 
 // SS10 startup impulse
-extern uint8_t start_up;
+extern uint16_t start_up;
 
 // SS11 seconds since power up
 extern uint32_t seconds;
@@ -55,16 +55,16 @@ extern uint32_t seconds;
 extern uint32_t minutes;
 
 // SS13 cluster status (not used)
-extern uint8_t cluster_state;
+extern uint16_t cluster_state;
 
 // SS14 telemetry status
-extern uint8_t telemetry_state;
+extern uint16_t telemetry_state;
 
 // SS15 cluster number
-extern uint8_t cluster_num;
+extern uint16_t cluster_num;
 
 // SS16..SS23 clusters online offline
-extern uint8_t cluster_link[8];
+extern uint16_t cluster_link[8];
 
 
 uint16_t used_ai = 0x0000;
